design: Show label always, Apply material 3, Correct view model file
